New cvmfs_repositories parameter
Normally `CVMFS_REPOSITORIES` in `/etc/cvmfs/default.local` is
populated from explicit cvmfs::mount instances depending on
the cvmfs_repo_list boolean per mount.

If `cvmfs_repositories` is specified then all values of `cvmfs_repo_list`
will be ignored and the CVMFS_REPOSITORIES will be exactly set to the
value of the new parameter.

This is useful when wanting to control which repositories will be probed
with cvmfs_config probe for instance.