apt: ensure RSA signing key is refreshed instead of old DSA key
Upstream has an older dsa1024 signing key and a more recent rsa4096 signing key in place
which is used for new releases.
While both keys are provided via the same source URL, existing systems installed before
the RSA key was added would never fetch that key, as the DSA key never expires and only
that is refreshed via Puppet.

Ensuring the RSA key is refreshed instead of the DSA key ensures that both keys are present
and that the RSA key is added also for existing systems.