Support failover yum repositories.
It is now possible to specify the `repo_base` parameter
as array as well as a string.

```puppet
class{ 'cvmfs':
  base_url => ['http://cvmrepo.s3.cern.ch/cvmrepo/yum/','http://cvmrepo.web.cern.ch/cvmrepo/yum']
}
```

in which both yum repositories will be used in failover.

A second yum repository has been configured as a part of default
repositories for yum installations.

Note that when using Debian family of OS there will no failover and the
first element of the array will be chosen if provided.